bench(h5): split online attack@k analyzer into a sibling module; scan code fences without a lazy regex
